PROGRESS 01-12-2024
The renderer is now working fully and has no bugs with my small tests in Program.cs (which are currently commented out).

Next up, I would like to optimize the StyleCode processing (params arrays are inefficient) as well as add screen resizing support. This should all be fairly trivial.

What will NOT be so trivial is when I eventually create a UI layer to run on top of this renderer. I want the UI layer to follow a similar ideology to this renderer by only submitting writes for screen elements that have changed.

The renderer would optimize needless writes away anyways, but we can go even faster by avoiding the write call to begin with. This added speed is what I want the UI layer to strive for.