Progress 2/5/2024
I actually managed to make the diff algorithm work.
It's not particularly quick, but it does pass the sine wave tracing test with no artifacting.

It's also nice because the renderer, in its current state, does not actually need a central framebuffer. We just track what has been written and what needs to be written, and diff between those accordingly.