Progress 2/6/2024
Added a concurrency aspect to the Canvas.Flush() method.
The RenderModifiedPixels and PlatformWriteStdout functions (and all they rely on) can be executed concurrently while the client code and ReconcileFrames() continue executing.

This concurrency system works up to one frame in advance.
(enough to benefit from concurrency but not so much as to complete de-sync the two concurrent parts)

(This all improved renderer performance by a fair bit on my machine. ~6k FPS -> ~10k FPS)