More Perf
Implemented hash-caching (that's fun to say) for both Pixel and Color24.
This regained a fair bit of performance (2 - 2.5k FPS) which was lost when I combined diffing and rendering to both happen on the render thread.

Conceptually, there are a few ways to make this whole thing perform better.

I can implement further optimizations for the render thread (which I am running out of).

Or, I can try to shift the work balance more towards the main thread (which currently spends a lot of its time waiting for the render thread to finish).
This would probably be through trying to further maximize precaching (as with the hashes in this commit).