SOTR (State of the Repository) 3/20/2024
So, I think I have made this renderer as fast as I can assuming I only work on it for a reasonable amount of time (barring potential rewrites or radical design changes).

As it stands now, the renderer is almost entirely held up by the Write Thread - which means that it is now bottlenecked by the throughput limit of writing to stdout (by several tens-of-thousands of nanoseconds).
This is something we cannot help without writing a custom terminal/graphics renderer - which is beyond the scope of this project (though not something I am shutting the door on).

I think this is the signal to end my quest for optimization, or at least put it on hold.

For now, I am going to focus on refactoring and just generally cleaning things up so that I can now properly make this into its own library.

I also have to come up with a cool name for this renderer. So far, I'm thinking "dT" (short for delta-time) might be a good punchy and short name, but I am not 100% sold yet.